Grab query string from URLs
A query string value was only taken into account when the middleware
was used to generate a page, but it's possible that a query string
value may be passed as part of a URL when generating with a function
like `quick_publish`. This change takes that into account and
always separates the query string value from the path.

Signed-off-by: Michael Dippery <mdippery@gmail.com>